# BE/UserService/superapp-user-profile-service/swagger-spec.json
{"openapi":"3.0.0","paths":{"/v2/api-docs":{"get":{"operationId":"AppController_swaggerDocument","summary":"","description":"This is the main Hello Wold Get Endpoint.","parameters":[],"responses":{"200":{"description":"Successfull displayed `Hello World`."},"403":{"description":"Forbidden."}}}},"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user/create":{"post":{"operationId":"UserController_regUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"201":{"description":""}}}},"/user/{msisdn}":{"get":{"operationId":"UserController_getUser","parameters":[],"responses":{"200":{"description":""}}}},"/user/update/{id}":{"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}}},"responses":{"200":{"description":""}}}},"/user/authenticate":{"post":{"operationId":"UserController_authenticateUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDTO"}}}},"responses":{"201":{"description":""}}}},"/user/logout":{"post":{"operationId":"UserController_logOutUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMobileDTO"}}}},"responses":{"201":{"description":""}}}},"/user/bonus":{"post":{"operationId":"UserController_claimBonus","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMobileDTO"}}}},"responses":{"201":{"description":""}}}},"/otp/send":{"post":{"operationId":"OTPController_createOTP","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OTPRequestDTO"}}}},"responses":{"201":{"description":""}}}}},"info":{"title":"User Service","description":"Super App user service","version":"1.0","contact":{"name":"Digital Telco Enabler","url":"http://www.axiatadigitallabs.com/","email":"dte@axiatadigitallabs.com"},"termsOfService":"Terms of services","license":{"name":"License of API","url":"API license URL"}},"tags":[],"servers":[],"components":{"schemas":{"UserDTO":{"type":"object","properties":{}},"UserUpdateDTO":{"type":"object","properties":{}},"AuthDTO":{"type":"object","properties":{}},"UserMobileDTO":{"type":"object","properties":{}},"OTPRequestDTO":{"type":"object","properties":{}}}}}